requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
tortoise-orm==0.20.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
aiosqlite<=0.18.0
pydantic==2.4.2
pydantic-settings==2.0.3
python-dotenv==1.0.0
aerich==0.7.2
APScheduler==3.10.4
bcrypt==4.0.1